+++ main_loop_helper.h
@@ -8,22 +8,22 @@
 
 class MainLoopHelper
 {
-    bool display_region_rect = false;
-    sf::RectangleShape region_selection_rect;
-    sf::RenderWindow *window;
+  bool display_region_rect = false;
+  sf::RectangleShape region_selection_rect;
+  sf::RenderWindow *window;
 
-    void initialize_auxiliary_entities();
+  void initialize_auxiliary_entities();
 
-  public:
-    MainLoopHelper(sf::RenderWindow *window);
-    ~MainLoopHelper();
-    void displayAuxiliaryEntities();
-    void setDisplayRectStart(sf::Vector2f pos);
-    void setDisplayRectEnd(sf::Vector2f pos);
-    Rectangle<float> getSelectedRegion();
-    void togleRegionRect(bool state);
+public:
+  MainLoopHelper(sf::RenderWindow *window);
+  ~MainLoopHelper();
+  void displayAuxiliaryEntities();
+  void setDisplayRectStart(sf::Vector2f pos);
+  void setDisplayRectEnd(sf::Vector2f pos);
+  Rectangle<unsigned int> getSelectedRegion();
+  void togleRegionRect(bool state);
 
-    bool isRegionDisplayed();
+  bool isRegionDisplayed();
 };
 
 #endif //HELPER_H
